requirements.txt
# Core data and analysis
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
pandas-ta>=0.3.14b0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0
mplfinance>=0.12.9b7

# AI and NLP
openai>=1.3.0
langchain>=0.0.340
langchain-openai>=0.0.2

# API and web
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Storage
python-dotenv>=1.0.0
